Remove wpnp-init.php and replace some of the functionality in install.php
This script caused a lot of headaches over time. We now…
a) assume that https is working and that's what we want. No more inexplicit http!
b) set defaults only on our first install routine, not on every page view
c) define settings for autoupdating in the wp-config.php

As a result, wpnp-init.php can be deleted.